refactor(app): clear wave-4 app test size allowlist via map suite split
Extract shared map widget tester/cache helpers and move trailing
ct_region_map pins into part4 so all app/test files sit under the
1000-line cap with an empty shrink-only allowlist.
