+++ app/test/ct_region_map_widget_part2_test.dart
@@ -20,7 +20,6 @@ import 'package:colonizethis_app/features/game/flame/region_map/region_map.dart'
     show
         BaseLayerDisplayMode,
         CtMapVisibilityMode,
-        CtRegionMapComponent,
         extractionIndicatorDisplaySizePx,
         extractionIndicatorRectsForIconRect,
         isCellUnderFleetRevealHalo,
@@ -37,39 +36,17 @@ import 'package:colonizethis_app/features/game/flame/region_map/region_map.dart'
         shouldApplyFogToLandBase,
         shouldWrapProvinceLabelPresenceIcons,
         visibilityForTerrainForMapCell;
-import 'package:colonizethis_app/features/game/flame/caches/civilian_icon_cache.dart';
-import 'package:colonizethis_app/features/game/flame/caches/province_label_icon_cache.dart';
 import 'package:colonizethis_app/features/game/flame/tilesets/tilesets.dart';
-import 'package:colonizethis_app/features/game/flame/caches/town_icon_cache.dart';
-import 'package:colonizethis_app/features/game/flame/region_map/ct_region_map_game.dart';
 import 'package:colonizethis_app/widgets/ct_region_map.dart' show CtRegionMap;
 
 import 'ct_region_map_test_support.dart';
 
-CtRegionMapComponent ctRegionMapComponentFromTester(WidgetTester tester) {
-  final finder = find.byWidgetPredicate(
-    (w) => w.runtimeType.toString().startsWith('GameWidget<'),
-  );
-  expect(finder, findsOneWidget);
-  final gameWidget = tester.widget(finder);
-  final game = (gameWidget as dynamic).game as CtRegionMapGame;
-  return game.debugMapComponentForTest;
-}
 
 void main() {
   suppressLogsForTests();
 
   group('CtRegionMap (Flame map widget)', () {
-    setUpAll(() async {
-      // CtRegionMapComponent.onLoad awaits these; without a warm cache, a single
-      // pump() is not enough when tests run alone (e.g. CI --total-shards).
-      await terrainTilesetCache.load();
-      await transportOverlayTilesetCache.load();
-      await resourceIconCache.load();
-      await civilianIconCache.load();
-      await townIconCache.load();
-      await provinceLabelIconCache.load();
-    });
+    setUpAll(warmCtRegionMapCachesForTests);
 
     testWidgets(
       'camera resize logic runs when parent size changes',
@@ -964,158 +941,6 @@ void main() {
         }
       },
       timeout: const Timeout(Duration(seconds: 12)),
-    );
-
-    testWidgets(
-      'map renders without resource icons in terrainOnly mode (SPEC/ui/map-widget.md § Base layer display mode)',
-      (WidgetTester tester) async {
-        await tester.runAsync(() async {
-          await terrainTilesetCache.load();
-        });
-
-        final region = ctRegionMapTestOldWorldRegion();
-        await tester.pumpWidget(
-          ctRegionMapTestHarness(
-            region: region,
-            baseLayerDisplayMode: BaseLayerDisplayMode.terrainOnly,
-          ),
-        );
-        await tester.pump();
-
-        // Widget should render without errors even without resource icons loaded
-        expect(find.byType(CtRegionMap), findsOneWidget);
-      },
-      timeout: const Timeout(Duration(seconds: 10)),
-    );
-
-    testWidgets(
-      'resource icons are fogged in player-constrained visibility mode for fogged tiles',
-      (WidgetTester tester) async {
-        await tester.runAsync(() async {
-          await terrainTilesetCache.load();
-          await resourceIconCache.load();
-        });
-
-        final base = ctRegionMapTestOldWorldRegion();
-        // Create a region with some fogged cells
-        final foggedCells = base.cells.map((c) {
-          if (c.isSea) return c;
-          final visibility = c.x < 2
-              ? TileVisibility.fogged
-              : TileVisibility.visible;
-          return CellViewData(
-            x: c.x,
-            y: c.y,
-            regionCellId: c.regionCellId,
-            isSea: c.isSea,
-            terrainType: c.terrainType,
-            resourceId: c.resourceId,
-            improvementLevel: c.improvementLevel,
-            roadLevel: c.roadLevel,
-            visibility: visibility,
-            ownerFactionId: c.ownerFactionId,
-          );
-        }).toList();
-
-        final region = RegionMapViewData(
-          regionId: base.regionId,
-          width: base.width,
-          height: base.height,
-          cellSize: base.cellSize,
-          cells: foggedCells,
-          capitalMarkers: base.capitalMarkers,
-          portMarkers: base.portMarkers,
-          factionColors: base.factionColors,
-          greatPowerFactionIds: base.greatPowerFactionIds,
-          terrainColors: base.terrainColors,
-          unitMarkers: base.unitMarkers,
-        );
-
-        await tester.pumpWidget(
-          ctRegionMapTestHarness(
-            region: region,
-            visibilityMode: CtMapVisibilityMode.playerConstrained,
-            playerViewForResources: ctRegionMapTestPlayerView,
-            baseLayerDisplayMode: BaseLayerDisplayMode.terrainAndResources,
-          ),
-        );
-        await tester.pump();
-
-        // Widget should render without errors
-        expect(find.byType(CtRegionMap), findsOneWidget);
-      },
-      timeout: const Timeout(Duration(seconds: 10)),
-    );
-
-    testWidgets(
-      'resource icons render for 64px tiles with quarter-size display (SPEC/ui/map-widget.md § Resource Icons)',
-      (WidgetTester tester) async {
-        await tester.runAsync(() async {
-          await terrainTilesetCache.load();
-          await resourceIconCache.load();
-        });
-
-        expect(resourceIconDisplaySizePx(64), equals(16));
-
-        final region = ctRegionMapTestOldWorldRegion();
-        await tester.pumpWidget(
-          ctRegionMapTestHarness(
-            region: region,
-            cellSizePx: 64,
-            baseLayerDisplayMode: BaseLayerDisplayMode.terrainAndResources,
-          ),
-        );
-        await tester.pump();
-
-        expect(find.byType(CtRegionMap), findsOneWidget);
-      },
-      timeout: const Timeout(Duration(seconds: 10)),
-    );
-
-    testWidgets(
-      'map renders correctly with 64px tile size (SPEC/ui/map-widget.md § Resource Icons)',
-      (WidgetTester tester) async {
-        await tester.runAsync(() async {
-          await terrainTilesetCache.load();
-          await resourceIconCache.load();
-        });
-
-        final region = ctRegionMapTestOldWorldRegion();
-        await tester.pumpWidget(
-          ctRegionMapTestHarness(
-            region: region,
-            cellSizePx: 64,
-            baseLayerDisplayMode: BaseLayerDisplayMode.terrainAndResources,
-          ),
-        );
-        await tester.pump();
-
-        expect(find.byType(CtRegionMap), findsOneWidget);
-      },
-      timeout: const Timeout(Duration(seconds: 10)),
-    );
-
-    testWidgets(
-      'map renders correctly with 16px tile size (SPEC/ui/map-widget.md § Resource Icons)',
-      (WidgetTester tester) async {
-        await tester.runAsync(() async {
-          await terrainTilesetCache.load();
-          await resourceIconCache.load();
-        });
-
-        final region = ctRegionMapTestOldWorldRegion();
-        await tester.pumpWidget(
-          ctRegionMapTestHarness(
-            region: region,
-            cellSizePx: 16,
-            baseLayerDisplayMode: BaseLayerDisplayMode.terrainAndResources,
-          ),
-        );
-        await tester.pump();
-
-        expect(find.byType(CtRegionMap), findsOneWidget);
-      },
-      timeout: const Timeout(Duration(seconds: 10)),
     );
   });
 }
